companion/requirements.txt
# ESO Build Optimizer Companion App Dependencies

# File watching
watchdog>=3.0.0

# HTTP client
httpx>=0.25.0

# System tray (optional, for GUI mode)
pystray>=0.19.0
Pillow>=10.0.0

# Packaging
pyinstaller>=6.0.0
